+++ ab2d-events-client/build.gradle
@@ -1,4 +1,4 @@
-version "$eventLibsVersion"
+version "$eventClientVersion"
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
