+++ ab2d-events-client/src/main/java/gov/cms/ab2d/eventclient/events/FileEvent.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.eventlibs.events;
+package gov.cms.ab2d.eventclient.events;
 
 
 import java.io.File;
