+++ ab2d-events-client/src/main/java/gov/cms/ab2d/eventclient/sqs/SendSQSEvent.java
@@ -1,15 +1,15 @@
-package gov.cms.ab2d.eventlibs.sqs;
+package gov.cms.ab2d.eventclient.sqs;
 
 import com.amazonaws.services.sqs.AmazonSQS;
 import com.amazonaws.services.sqs.model.SendMessageRequest;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import gov.cms.ab2d.eventlibs.events.LoggableEvent;
+import gov.cms.ab2d.eventclient.events.LoggableEvent;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 
-import static gov.cms.ab2d.eventlibs.sqs.SQSConfig.EVENTS_QUEUE;
+import static gov.cms.ab2d.eventclient.sqs.SQSConfig.EVENTS_QUEUE;
 
 @Slf4j
 @Component
