+++ ab2d-events-client/src/test/java/gov/cms/ab2d/eventclient/SpringBootApp.java
@@ -1,10 +1,10 @@
-package gov.cms.ab2d.eventlibs;
+package gov.cms.ab2d.eventclient;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication(
-        scanBasePackages = {"gov.cms.ab2d.eventlibs"},
+        scanBasePackages = {"gov.cms.ab2d.eventclient"},
         exclude = {
                 org.springframework.cloud.aws.autoconfigure.context.ContextInstanceDataAutoConfiguration.class,
                 org.springframework.cloud.aws.autoconfigure.context.ContextStackAutoConfiguration.class,
