+++ ab2d-events-client/src/test/java/gov/cms/ab2d/eventclient/config/Ab2dEnvironmentTest.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.eventlibs.config;
+package gov.cms.ab2d.eventclient.config;
 
 
 import org.junit.jupiter.api.Test;
