+++ ab2d-events-client/src/test/java/gov/cms/ab2d/eventclient/events/JobStatusChangeEventTest.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.eventlibs.events;
+package gov.cms.ab2d.eventclient.events;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
