+++ build.gradle
@@ -22,7 +22,7 @@ ext {
     bfdVersion='1.2'
     filtersVersion='1.5'
     commonsLangVersion='3.12.0'
-    eventLibsVersion='1.0'
+    eventClientVersion='1.0'
 
     sourcesRepo = 'ab2d-maven-repo'
     deployerRepo = 'ab2d-main'
